Update methods to reflect platform information
Hi there!!

This PR updates the implementations of the `type` and `platform` methods to reflect more accurate system information.

I updated the implementation of `type` to simply return "React Native", and the implementation of `platform` to return the current platform, either `android`, or `ios`, or whatever else might be returned by `Platform.OS`.

I'm guessing this is going to break someone's something, but I think it more accurately reflects the `os` in question. I can't see how this module would be run in a browser environment, and as such I think it's misleading and inaccurate to return `browser` for those methods. I'd love to have a discussion around this if you disagree with my changes. Please let me know!

Thanks again!